My next commit. Added fonts, some colors, trying to make it look decent by 6.
The jumbotron is brainfucking me because its height is impossible to control with a percentage.

Anyways...

Debating on whether or not to make the side nav items into buttons. Either way they could use some better effects and animations.

Need to clean up the site as much as possible before they can see it.